Move from Int64 to Int, IrisWebsocketService: Update docs, IrisWebsocketServiceImpl: Add Task.isCancelled check to avoid dangling streams

+++ ArtemisKit/Sources/Iris/Models/IrisCourseSettings.swift
@@ -35,7 +35,7 @@ struct IrisCourseSettingsDTO: Codable, Hashable {
 /// Full settings response from `GET /courses/{courseId}/iris-settings`.
 /// Combines the editable settings with the server-computed effective limits.
 struct IrisCourseSettingsWithRateLimitDTO: Codable, Hashable {
-    let courseId: Int64
+    let courseId: Int
     let settings: IrisCourseSettingsDTO
     let effectiveRateLimit: IrisRateLimitConfiguration
     let applicationRateLimitDefaults: IrisRateLimitConfiguration
