+++ ArtemisKit/Sources/Iris/Services/IrisChatService/IrisChatHttpService.swift
@@ -8,21 +8,22 @@
 import Common
 
 protocol IrisChatHttpService {
-    func getCurrentOrCreateSession(mode: ChatServiceMode, entityId: Int64) async -> DataState<IrisSession>
-    func createSession(mode: ChatServiceMode, entityId: Int64) async -> DataState<IrisSession>
+    func getCurrentOrCreateSession(mode: ChatServiceMode, entityId: Int) async -> DataState<IrisSession>
+    func createSession(mode: ChatServiceMode, entityId: Int) async -> DataState<IrisSession>
 
+    //TODO: courseId Long ?
     func getChatSessions(courseId: Int) async -> DataState<[IrisSessionDTO]>
-    func getChatSession(courseId: Int, sessionId: Int64) async -> DataState<IrisSession>
+    func getChatSession(courseId: Int, sessionId: Int) async -> DataState<IrisSession>
 
     func getSessionAndMessageCount() async -> DataState<IrisSessionCountDTO>
 
     func deleteAllSessions() async -> NetworkResponse
-    func deleteSession(sessionId: Int64) async -> NetworkResponse
+    func deleteSession(sessionId: Int) async -> NetworkResponse
 
-    func getMessages(sessionId: Int64) async -> DataState<[IrisMessageResponseDTO]>
-    func sendMessage(sessionId: Int64, request: IrisMessageRequestDTO) async -> DataState<IrisMessageResponseDTO>
-    func resendMessage(sessionId: Int64, messageId: Int64) async -> DataState<IrisMessageResponseDTO>
-    func rateMessage(sessionId: Int64, messageId: Int64, helpful: Bool) async -> DataState<IrisMessageResponseDTO>
+    func getMessages(sessionId: Int) async -> DataState<[IrisMessageResponseDTO]>
+    func sendMessage(sessionId: Int, request: IrisMessageRequestDTO) async -> DataState<IrisMessageResponseDTO>
+    func resendMessage(sessionId: Int, messageId: Int) async -> DataState<IrisMessageResponseDTO>
+    func rateMessage(sessionId: Int, messageId: Int, helpful: Bool) async -> DataState<IrisMessageResponseDTO>
 }
 
 enum IrisChatHttpServiceFactory: DependencyFactory {
